requirements.txt
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
natsort==8.4.0 